Slideshow app Testing using Yeoman and AngularJs
Testing this app for setup and build out of code using Yeoman and
AngularJS. The main focus is to use JSON files to create a scalable
index for images and the details that will accompany them. Creating a
lightweight app that can be scalable for slideshows to Portfolio sites
that are image heavy.